Updating minor changes in ed25519.
