+++ sign/ed25519/modular_test.go
@@ -2,7 +2,6 @@ package ed25519
 
 import (
 	"crypto/rand"
-	"math/big"
 	"testing"
 
 	"github.com/cloudflare/circl/internal/conv"
@@ -15,8 +14,7 @@ func TestCalculateS(t *testing.T) {
 	k := make([]byte, Size)
 	r := make([]byte, Size)
 	a := make([]byte, Size)
-	var order big.Int
-	order.SetString("1000000000000000000000000000000014def9dea2f79cd65812631a5cf5d3ed", 16)
+	orderBig := conv.BytesLe2BigInt(order[:])
 
 	for i := 0; i < testTimes; i++ {
 		_, _ = rand.Read(k[:])
@@ -30,7 +28,7 @@ func TestCalculateS(t *testing.T) {
 		got := conv.BytesLe2BigInt(s[:])
 
 		bigK.Mul(bigK, bigA).Add(bigK, bigR)
-		want := bigK.Mod(bigK, &order)
+		want := bigK.Mod(bigK, orderBig)
 
 		if got.Cmp(want) != 0 {
 			test.ReportError(t, got, want, k, r, a)
@@ -41,8 +39,7 @@ func TestCalculateS(t *testing.T) {
 func TestReduction(t *testing.T) {
 	const testTimes = 1 << 10
 	var x, y [Size * 2]byte
-	var order big.Int
-	order.SetString("1000000000000000000000000000000014def9dea2f79cd65812631a5cf5d3ed", 16)
+	orderBig := conv.BytesLe2BigInt(order[:])
 
 	for i := 0; i < testTimes; i++ {
 		for _, j := range []int{Size, 2 * Size} {
@@ -53,7 +50,7 @@ func TestReduction(t *testing.T) {
 			reduceModOrder(y[:j], true)
 			got := conv.BytesLe2BigInt(y[:])
 
-			want := bigX.Mod(bigX, &order)
+			want := bigX.Mod(bigX, orderBig)
 
 			if got.Cmp(want) != 0 {
 				test.ReportError(t, got, want, x)
